+++ CHANGELOG.md
@@ -5,9 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Year-by-year temperature chart. The old temperature chart is selectable through a drop-down
+* Selection of min, mean and max temperature through switches in year-by-year chart
+* Add pan and zoom in the x direction through mouse drag and mouse wheel to year-by-year temperature chart
+* Add pan and zoom in the x direction through mouse drag and mouse wheel to temperature chart
+* Color offset selection through slider to year-by-year temperature chart
+* Selection of points and lines to year-by-year temperature chart
+* Persistence of year-by-year temperature chart controls in browser local storage
+
 ### Changed
 * Rename page to Emil's Weather Station @ bergetvidhandfatet
 * Change favicon to cloudy weater
+* Rework using Chart.js
+* Use unique colors in year-by-year temperature chart
 
 ### Fixed
 * Update bearer token